Moved @draw_on_return and draw_shape() from primitives.py to primitves3d.py
