+++ p5/core/primitives3d.py
@@ -17,11 +17,10 @@
 #
 
 import math
-
+import functools
 from .geometry import Geometry
-
+from . import p5
 from ..pmath import matrix
-from .primitives import _draw_on_return
 
 # We use these in ellipse tessellation. The algorithm is similar to
 # the one used in Processing and the we compute the number of
@@ -44,6 +43,39 @@ from .primitives import _draw_on_return
 MIN_POINT_ACCURACY = 20
 MAX_POINT_ACCURACY = 200
 POINT_ACCURACY_FACTOR = 10
+
+
+def _draw_on_return(func):
+    """Set shape parameters to default renderer parameters
+
+    """
+
+    @functools.wraps(func)
+    def wrapped(*args, **kwargs):
+        s = func(*args, **kwargs)
+        draw_shape(s)
+        return s
+
+    return wrapped
+
+
+def draw_shape(shape, pos=(0, 0, 0)):
+    """Draw the given shape at the specified location.
+
+    :param shape: The shape that needs to be drawn.
+    :type shape: p5.PShape
+
+    :param pos: Position of the shape
+    :type pos: tuple | Vector
+
+    """
+    p5.renderer.render(shape)
+
+    if isinstance(shape, Geometry):
+        return
+
+    for child_shape in shape.children:
+        draw_shape(child_shape)
 
 
 @_draw_on_return
